Add a test script for RCS support.
Some bits are untested; the reason is that 'vcs up' and 'vcs rm' don't
work quite the same as for the other systems.

First 'vcs rm' deletes immediately.  This could be fixed by the same
technique as 'vcs add'.

Secondly 'vcs up' will only make sure missing working copies are
present.  It will not merge changes in ,v files into working copies.
This is because it has no idea which version was checked out so it
cannot infer a base for the merge.  vcs could record which version was
checked out but that does not help if the user (or some other front
end) invokes rcs commands directly - the record might be missing or
wrong.  So this limitation is not likely to be fixable.